Added Dialog.
